ZAPP-196 Add routing to a new cluster specific view, move some code that happens on mount to app.vue instead of zcllayout
